* Fixes #32   You can specify via <configuration> in the pom, the following now:
  <httpPort>8080</httpPort>
  <portOffset>0</portOffset>
  <bindAddress>0.0.0.0</bindAddress>

Those are the defaults if not specified.

If they are specified, they are used both for `mvn wildfly-swarm:run`
and are baked into the -swarm.jar fat-jar as the defaults.

When running the fatjar, they can still be overridden by using the
appropriate -Djboss.* properties, such as:

  -Djboss.http.port=9191
  -Djboss.bind.adddress=127.0.0.2
  -Djboss.socket.binding.port-offset=2

* Fixes #48

Consistently use GAV'd names for the project artifact, instead of
sometimes GAV, sometimes finalName.