Filter invalid files
Motivation
SWARM-256. Windows build fails in Arquillian test because the project build
directory starts with 'c:\' which means that split(':') gives us a lone 'c'.

Modifications:
Filter out of the stream any files that don't exist on the filesystem before
all results are collected.

Result:
'c' results in a non existant file, so is filtered out. The original '\modules'
path that we wanted to add is still added successfully