Break apart Ribbon from Topology-management.
Ribbon normally uses Eureka, but initially we replaced it with
JGroups, for service discovery.  This ended up being sub-par, as
there are other service discovery mechanisms (such as Eureka, or
Consul) which users may want to use.

By dividing the Ribbon and Topology bits, all of our Ribbon stuff
is topology-infrastructure agnostic, and now has a Topology abstraction
between itself and JGroups or other mechanisms.

This has introduced a 'topology' fraction which Ribbon talks to
on one side, and begins specific implementations (such as topology-jgroups)
to populate information from the other side.

Another benefit of this is that topology facilities may be used
without involving Ribbon at all now.

Additionaly, move ribbon-webapp to topology-webapp, along with
renaming the bits to /topology/topology.js and /topology/system/stream.