Don't resolve artifacts that exist locally
Motivation:
We resolve all dependencies when building a swarmjar that doesn't
bundleDependencies to ensure they are in the local M2_REPO. For most
dependencies, the artifacts are already there, so this is just wasted
time.

Modifications:
* check the local repository for the artifact before trying remote
  resolution

Result:
This results in a ~45% improvement in the local build time for a full
build (34 min -> 19 min).