Subject: Code styles
Motivation:
Update existing files of the project to conform to previously defined code styles.

Modifications:
Most of the changes are within pom.xml files. There are a few changes in Java files to change the order of variables and methods, and some spacing.

Result:
No impact on the product as it's simply a modification to code position