Make swagger a zero-configuration fraction.
Motivation:
Developers should be able to include the swagger fraction in a JAX-RS
application without having to do much of anything besides add the
swagger annotations to their resources. Prior to this change, use of the
Swagger fraction meant that you had to use ShrinkWrap to deploy your app.

Modifications:
When a swagger-enabled application is deployed, the SwaggerConfiguration
instance checks to see if any packages have been configured for scanning
by the Swagger scanner. If not, it will scan the deployment for the first
package it finds that includes a .class file, excluding everything in
the org.wildfly.swarm package space.

Result:
Now a swagger-enabled application can be explicitly configured, or work
in a reasonable way with zero configuration. Win.