Don't strip WildFly Swarm dependencies from WARs
Motivation:
We try to remove any swarm dependencies (and their transitives) from WAR
files when they pass through BuildTool in an effort to reduce file
size. The mechanism for doing this is overly complex, and increases the
build time since it requires a second dependency resolution step. It
also isn't used when testing with arquillian or when using StartMojo.

The proper solution here is to not require swarm dependencies for
WAR-based builds, and instead auto-detect what fractions are
needed (coming soon).

Modifications:
* remove dependency stripping

Result:
Dependencies are left alone.