Enhance fraction detection to be more flexible
Motivation:
We have some fractions (e.g. jaxrs-weld) that should only be activated
if two packages are being used. We have other fractions (e.g. spring)
that needs to be activated if any org.springframework.* packages are in
use. The detection mechanism needs to be improved to handle those cases.

Modifications:
* allow for anding (with +) packages
* allow for splat matching (with *) packages
* add missing fractions to the fraction-packages list

Result:
More fractions can be detected.

Concerns:
The current implementation requires all detectable fractions to be
listed in fraction-packages.properties, even if they aren't in
core (swagger, spring, etc). We may need to come up with a different way
to specify these.