Adjust FractionList to implement iface and provide package specs
Motivation:
Since we now use FractionList in core/tools/, the implementation of
FractionList needs to implement the core/tools/FractionList
interface. The interface expands the scope of FractionList to provide
package specs for fraction detection, so this needs to be implemented as
well.

Modifications:
* remove FractionDescriptor, since it is now provided by core/tools/
* implement the tools/FractionList interface
* provide package specs from FractionDescriptor

Result:
The FractionList implementation can now be used by
core/tools/BuildTool.