Don't try to add bootstrap as fraction
Motivation:
FractionUsageAnalyzer tries to add base fractions to the list of
discovered fractions so that the bare minimum is always available. It
was adding both container and bootstrap, but the latter isn't a
fraction, so resulted in a null in the resolved list since FractionList
knows nothing about bootstrap. We don't need to depend directly on
bootstrap anyway, since container will bring it in.

Modifications:
* remove bootstrap from the default fraction set
* simplify the code that adds the defaults, since there is now only one

Result:
FractionUsageAnalyzer now gives correct, usable results.