Remove a variety of FooConfiguration implementations.
The SPI package, as of Beta5, has two annotations that
may be applied to FooFraction in order to obviate the
need for authoring a FooConfiguration implementation.

At the class-level of a given Fraction:

@Configuration(
  marshal=true,
  extension="org.jboss.as.fancypants",
  ignorable=false,
  parserFactoryClassName="org.wildfly.swarm.fancy.runtime.FancyParserFactory"
)

If the fraction does not provide a public no-arg constructor
for creating the default fraction, exactly 1 public no-arg
static method may be annotated with

@Default
public static FancyFraction createMyFancyFraction() {...}

And that method will be used to provide the default fraction.