Subject: Alter addAllDependencies() behavior
Motivation:
SWARM-389, we want to prevent the default behavior of Archive.addAllDependencies() including WildFly Swarm dependencies in WEB-INF/lib as they are already available via JBoss Modules. Not excluding them means anyone wanting to create a custom Archive in a main() would need to use Archive.addDependency() for every external dependency because addAllDependencies() would break the application.

Modifications:
Modify addAllDependencies() to pass 'org.wildfly.swarm' as an exclusion filter to ArtifactLookup.allArtifacts() by default, but provide addAllDependencies(true) as a way to include WildFly Swarm dependencies if you so choose

Result:
Should prevent ClassCastExceptions in a users application when addAllDependencies() is used in a custom main()