Subject: Reload static content without needing custom Main
Motivation:
This commit fixes SWARM-230 by ensuring static content changes get
reflected when developing from an IDE even if the user does not
specify a custom Main class.

Modifications:
The Archive returned from DefaultWarDeploymentFactory#create now calls
the StaticContentContainer#staticContent() method by default. As a
side-effect of this change, UndertowExternalMountsAsset had to be
modified to be able to deserialize itself like our other Asset
implementations already do.

Result:
Making a change to static content when running in an IDE without a
custom Main is now immediately reflected when refreshing the
browser. The integration test for the static-war example in the
wildfly-swarm-examples repo will be updated to test this functionality
in a separate commit.