Create Jandex indexes for ribbon-secured
When the change was made to create indices for all of our fractions,
it looks like ribbon-secured was left out. This fixes that and adds a
test to verify that the ribbon secured client API can actually be used
from within web code, and thus the RibbonSecuredConfiguration was
found via the index.