Use the latest fraction-plugin for simple fractions.
New plugin allows for <swarm.fraction.bootstrap>true</swarm.fraction.bootstrap>
in the pom.xml of modules-only fractions, to place the marker file there.  So we
no longer need to stick on in src/main/resources/

All markers removed were empty, and re-created as empty via the plugin.  Basicly
a no-op, but a fancy no-op which allows deleting files.