Fix for SWARM-189 - Do not enable HTTP listeners until deployment.
Prior to this, when the container was start()'d, undertow would
immediately start listening with zero deployments.  This would
cause the HTTP endpoint to be open, accepting connections, but
obviously not connecting them to any particular application.

Now, the UndertowFraction by default produces listeners that
are enabled=false.  When a container is start()'d, undertow
is setup and running, but without any actual HTTP listeners
ready to accept connections, resulting in "connection refused"
until a .war (or .ear) is deployed.

Upon successful deployment of a .war (or .ear), all http listeners are
enabled, hopefully providing a consistent response from the
deployed application, instead of the in-between state of having
the web-server up but the application not responsive.

Container added a .start(bool) method to eagerly start the
listeners prior to deployments, mostly for use by the MonitorFraction's
own tests, which want to test system-wide monitors not related
to any application.