Further work towards SWARM-189.
Seems that JAX-WS (SOAP) doesn't like to start deployments
if there is no listener.  So instead while we start with
the listeners in enabled=false (MSC mode NEVER), before deployment
we move them to MSC mode ON_DEMAND, which satisfies JAXWS.

It also seems that JAX-RS doesn't create enough demand to turn
the ON_DEMAND to ACTIVE, so we continue to perform that action
after deployment, still, also.