Subject: JAX-RS Exception Mappers
Motivation:
SWARM-327, JAX-RS exception mapper in a user application was never being triggered. User exception mapper was blocked because we had created an exception mapper to prevent an exception in the log when favicon.ico was requested but not present in a user application.

Need to modify how we prevent that exception so that a user can create their own exception mapper for javax.ws.rs.NotFoundException.

Modifications:
Remove old favicon.ico exception mappers and add an Undertow ServletExtension into the deployment that adds an HttpHandler to the chain to directly handle favicon.ico requests.

Result:
No change in how favicon.ico is handled, but users can now create their own JAX-RS exception mapper for NotFoundException