Subject: Multi module project in IDE
Motivation:
SWARM-383, when executing the main() of a multi module project in an IDE the target/classes of dependent modules are added to the Archive, but without a .jar extension. This results in them not being processed as dependent libraries by WF subsystems

Modifications:
Explicitly set a name for the Archive library being added by the ArtifactManager if we're dealing with a directory, such as target/classes.

Also prevent addAllDependencies() being called twice in the DefaultWarDeploymentFactory. It wasn't an issue per se, just added extra overhead that was not needed.

Result:
Multi module projects will now correctly load dependent modules from target/classes