Publish exposed-components.json from the top level [SWARM-409]
It's easy to forget to update exposed-components.json when it isn't at
the top level. The only reason it wasn't at the top level was to publish
it under a non-parent name, but we now publish the name of the module in
the json as well, so that doesn't matter.