Provide a way for an Arq test to provide a main().
Since the SpringBoot integration is slightly different
than typical WildFly Swarm usage, to effectively ARQ-test
a SpringBoot application, we need to specify a main() and
not rely upon a default deployment.

By adding a @WithMain annotation to the test class, you can
specify a class to use as the main() in the executable jar.

This annotation is currently only supported by the UberJar
container and not via InVm running.