Allow for fraction scanning of multiple sources [SWARM-411]
Motivation:
In order to be able to scan jars, we need to be able to scan the
dependency jars as well. This isn't an issue with wars, since the
dependencies are already included in the war, and the scanning is
recursive.

Modifications:
* expose a way to set multiple sources on the analyzer
* scan and combine all sources

Result:
Multiple sources can be scanned in one pass.