Add all -api to the exposed-components.json
This is in order to ensure all -api artifacts are also
present in the BOM so as to be dependencyManage'd.

Since the -api can transitively come in through other
bits, from a user's POV, they should be concurrently
managed along with -modules.

For instance, swagger-api brings in msc-api, and without
management, might get the wrong version, while swagger-modules
brings in msc-modules, which has been previously managed.

Sometimes you get skew and mismatch.  Now we hopefully don't.