Subject: Provided dependencies in addAllDependencies()
Motivation:
SWARM-412. When running in the IDE or `mvn wildfly-swarm:run` there are many jars added to the deployment with a call to Archive.addAllDependencies() that are actually provided within various modules. This often results in class cast or unable to link errors as the same class is found in WEB-INF/lib and a module. This fix will remove the provided libraries when not running in an uber jar

Modifications:
ArtifactManager.addAllDependencies() now reads WildFly Swarm jars for wildfly-swarm-classpath.conf to determine what should be provided, and excludes those jars from the resulting classpath

Result:
Should fix the problem of provided dependencies being added into an Archive when calling addAllDependencies() from a main()