Subject: Dependencies of `pom` type
Motivation:
SWARM-381. If a dependency of a user application has a `pom` type imported in the dependencies section, we try and add that `pom` as a `jar` dependency when Archive.addAllDependencies() is called.

`pom` is not a dependency we can deal with, so should be ignored.

Modifications:
ArtifactManager now ignores a classpath entry that ends with `.pom`.

Result:
A user application being run `mvn wildfly-swarm:run` will no longer crash if a project dependency imports a `pom`.