SWARM-415 - System property to eagerly open HTTP listeners.
Since introducing the delayed open HTTP listeners, it has been determined
that maybe it'd be useful to be able to eagerly open the listeners
through usage of a property.

System property swarm.http.eager, if non-null, will cause the
http listeners to not be lazy.

Rev SPI to include the newer SwarmProperty.