SWARM-463 - Allow partial configuration via standalone.xml
If a standalone.xml is present, it may configure all, some
or none of the container.

We now track which parts of standalone.xml are actually present
(at the <extension> level) within the provided .xml.

If configuration is performed for a particular subsystem
via the .xml, further configuration is inhibited through or
usual fraction defaults.

The big change is that previously, the presence of an XML config
inhibited *all* fraction defaults.  In the case of a partial
XML config, now, un-configured portions will be configured.

This provides for a hybrid approach to configuration, allowing
Swarm defaults to apply, but .XML to override or replace in
those instances where necessary.

Additionally, interface and socket-binding-groups allow
the .xml to win, or if unspecified, still use the normal
WFS defaults.