SWARM-464: Provide a warning (error, actually) for erroneous addPackage()
If a project is <packaging>war</packaging> and has a main(),
if the main() then does addPackage() on an internal synthetic
ShrinkWrap archive, it absolutely will not behave as a user
might imaging it would.

In this case, we throw an exception with a useful (hopefully)
description.

Currently JARArchive, WARArchive and JAXRSArchives are protected
thusly, for all flavors of addPackage() and addPackages().

This is accomplished by setting the default-deployment type
into a system property during the constructor of Container().

If addPackage(s) is called, this type is inspected, and if it's
.war, an exception is thrown.