SWARM-342: Enhance DatasourceArchive
Previously, a user wanting to add a datasource as a deployment
would have to create a JARArchive, and then .as(DatasourceArchive)
in order to then actually add one or more data-source config
trees to it.

The same pattern can be used, but a DatasourceArchive can also
now be directly created using ShrinkWrap.create(DatasourceArchive.class).

For instance:

  DatasourceArchive dsArchive = ShrinkWrap.create(DatasourceArchive.class);
  dsArchive.dataSource("ExampleDS", (ds) -> {
      ds.connectionUrl("jdbc:h2:mem:test;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE");
      ds.driverName("h2");
      ds.userName("sa");
      ds.password("sa");
  });