Don't allow boostrap from user-provided fractions to prevent auto-detection
The default auto-detect mode (when_missing) skips auto-detection if
bootstrap is found in the dependencies. If we add the user-provided
fractions (not ones specified as dependencies) to the dependencies
before looking for bootstrap, this will disable
auto-detection. User-provided fractions are intended to augment
auto-detection, and should be treated differently than
directly-dependent fractions.

This changes BuildTool to look for bootstrap before adding the user's
fractions (which would just look at actual dependencies). It then adds
the user's fractions, attempts auto-detection, then checks again for
bootstrap, at which point it should exist unless auto-detection was
disabled and no fractions were specified.