SWARM-473: Better CLI handling
This change adds to the default Swarm.main() support for command-line
arguments similar to those supported by WildFly proper.

This includes:

  -D<name>[=<value>]
  -P<url>
  -b <bind-address>
  -c <server-config>

Plus other fun things such as

  -s <stage-config>

The underly API is user-facing, and the user can use CommandLine directly
to parse the String...args provided to his main().  CommandLine provides
two methods which may (should) be used:

  cmdLine.applyProperties()

which applies -D, -P and -b arguments are system properties.

Along with

  cmdLine.applyConfigurations(container)

which applies the -c and -s configurations to the container.