SWARM-504 Support javaVmArguments in arquillian.xml
Using the typical arquillian.xml, you can now specify
configuration such as:

  <property name="javaVmArguments">-Dsome.property=a</property>

This will be passed to the arquillian container. When using the
default UberJarSimpleContainer, (without the @InVM annotation
on your test), then all javaVmArguments should be supported
(including -Xmx and such).

When using an InVmSimpleContainer, any argument that is not
used for setting properties (-Dfoo or -Dfoo=bar) will generate
a warning on the console, while applying any property argument
to the current JVM.