iQEcBAABCgAGBQJXaDrmAAoJEDZAhdNpr0OB1b8H/ir2MYYw/N9uUUWOJzcPZ3dM  us24phhzbuRf5uEC8x0SK34Bs/L1eBoyfecQ7lQ0roGKdhkbPzqBo2x2hTD97NtB  kfE7D1oKmt5rdBHL2KtF/EEv5Itvmct2/vjLqzcAu2W9oRnUkChGlfPGtSVZOyK+  Mad+/+aP3K7QP3yd1UMy1pUfbHbpmCiBk2RyxNw8FPh1VKaN3JrX4X3/38QYBEY/  oI5+CWpVeZkKEEHNsWNV/2TEPCE971FFND4hmnA5FjqXQSyRvZccq55FZFL1O872  qEhKkVdLIPCt2RQQuChqb9IgQq5ouGsFUE9Ch+REidVJd2SLd3A38qlAYPv7b8E=  =53GB  -----END PGP SIGNATURE-----
Use new javadocs repo instead of wildfly-swarm:gh-pages

Adding the docs to this repo for every build means everyone that pulls
wildfly-swarm has to download the updated docs if CI has built since
their last pull, which is time-consuming. Moving the docs out to a repo
that most devs will never need to pull solves that.