Fix for SWARM-530
Include the jackson-jaxrs-provider bits by default
through jaxrs.  If it is included through a user's
WEB-INF/lib/ via their own dependency, then its own
@Providers get registered and subvert ExceptionMappering
a variety of Jackson-specific exceptions.

Part and parcel to this, cause ArtifactManager, when
running tests, to be more aggressive for scanning for
wildfly-swarm-classpath.conf files.  In the 99% case,
it only scans things in the classpath that live in
your maven repository.

But...

In our own tests, when built from the root (or anywhere
higher than test/**), the artifacts such as jaxrs-api.jar
are NOT in the local maven repository, but rather in
jaxrs-api/target/jaxrs-api.jar.

The previous logic would not scan their wildfly-swarm-classpath.conf
file, causing the jaxrs-providers to be included when run from
a higher level.