Export the resteasy JAXB provider through the JAXRS fraction
To support "text/xml" and "application/json" encodings out of the box, we export the jaxb provider on the jaxrs runtime module.xml. This allows for straight forward usage of these contents types in service-service interactions and client-service testing in arq in-container tests.

The use cases where client-service interactions are tested with arq @RunAsClient need to import the relevant dependencies (i.e. reasteasy-client & providers) as <scope>test</scope>.