SWARM-599: Remote JMX not working.
Add dependencies to jmx fraction to include both
remoting and management, as that's what's required in order
for remote JMX to function.

Additionally, ensure that the correct remotingjmx module is
included into the solution of the module tree.

Add a test to verify that a remote client can indeed contact
and use the remote JMX interface.