SWARM-604 - Don't explode contents of uberjar repeatedly.
* Motivation:

When resolving the <artifact> entries of a module.xml, if multiple
asked or referred to the same artifact (because of our :api,
and :runtime module split), these jars were being exploded out
of the uberjar multiple times.

* Modifications:

The maven resolver used by the uberjar now caches the path of any
exploded jar in order to avoid duplicitous explosions.

* Result:

Less time exploding. Less space on disk with duplicated jars.