SWARM-598 - Allow selection of JMX endpoints without main()
* Motivation:

While we provide mechanisms for enabling and selecting a remote
JMX endpoint when a user provides a main(), we have no capability
for doing the same without a main(), such as in a vanilla .war.

* Modifications:

Added a swarm.jmx.remote configuration key, which can be set to

* 'management' to explicitly request the management interface.
* 'http' to request the standard-sockets http interface.
* any other non-null value to allow auto-selection.

* Result:

It should now be possible to enable remote JMX endpoint without
providing a main(), by using -Dswarm.jmx.remote or providing
a value through project-stages.yml.