SWARM-293 - Separate -swarm container from deployments.
* Motivation:

When dockering or other things, it may be nice to have the
runtime container separate from the deployment.  Container
could be placed on a docker layer, while the deployment is
updated, as a smaller item, on a higher layer.

* Modifications:

Added a <hollow>true</hollow> paramter to the package
Maven task, which if set, creates a deployment-less -swarm.jar.

This option passes to BuildTool#hollow(boolean) and thus may
also be easily exposed through the SwarmTool at some point.

Additionally, it wires up a HollowSwarm as the main(),
since there is no user-provided main().  This main iterates
all extra command-line arguments, creating ShrinkWrap
archives from them, and deploys them.  Anything deployable
may be accepted.

* Result:

If a project is built with <hollow>true</hollow>, then
the resulting -swarm.jar does not contain the deployment.