SWARM-616 - Remove extra jars from WEB-INF/lib
* Motivation:

When a user's .war is bundled in the -swarm.jar, depending on how
their build was configured, some WildFly Swarm fraction jars may
end up in their WEB-INF/lib.  When this occurs, particularly if
they are using CDI in their application, the fraction jars own
CDI components get mixed into their application.  This is not a
good thing.

* Modifications:

When building the -swarm.jar, we scan the user's own artifact for
any WEB-INF/lib/*.jar which contains wildfly-swarm-bootstrap.conf.
If a jar contains that marker, we remove it from the .war that
ultimately goes into the -swarm.jar.

This does still leave the user with a broken .war if they choose
to deploy it in a non-Swarm environment.

The best solution, probably, is to use auto-detect, but sometimes
this is not possible.

* Result:

Slightly cleaner .war files are embedded within the -swarm.jar.