SWARM-616 (again) - Remove gratuitous stuff from WEB-INF/lib
Motivation
----------

As previously, there's entirely too much cruft inside WEB-INF/lib
that shouldn't be there.

Modifications
-------------

We now track explicit top-level dependencies asked for by an
application, weed out the wildfly-swarm ones (via detection of
wildfly-swarm-bootstrap.conf), and recalculate the dependencies.

Result
------

This allows an app to ask for a jar that Swarm needs, and still
get it in its WEB-INF/lib.  But things only needed by Swarm proper
will get removed from WEB-INF/lib.

Additionally, if a .war is the primary artifact of a user's
application, we re-package it with a cleansed WEB-INF/lib,
leaving a .war.original in the build directory.