Subject: SWARM-596, Test class with inheritence or interfaces
Motivation
-----
Currently when @CreateSwarm is present on the method of a test class that has super types or implements interfaces it fails to load the class as these inherited types are not present in the Archive being classloaded.

Modifications
-----
Modify the UberJarSimpleContainer to recursively find all super types and interfaces to add the classes into the deployment archive.

Also modified the way the classes are added to utilize ClassContainer as its common to jar or war types but handles the correct /classes location depending on the archive type. There is an exception for ResourceAdapterArchive, but I believe it to be a bug in Shrinkwrap which has been raised: SHRINKWRAP-510

Result
-----
No impact on product, but allows test classes to inherit classes and implement interfaces when @CreateSwarm is annotated on a method