SWARM-628: Introduce configuration to generate self-signed certificate
Motivation
----------
This is more focused on quick-enabling HTTP/s for development purposes.
HTTP/2 support was introduced in SWARM-619 and is only enabled on HTTP/s listeners.

Modifications
-------------
Introduced two new configuration options: swarm.https.certificate.generate and swarm.https.certificate.generate.host

Result
------
When the new configuration option is set, it will generate a self-signed certificate and enable HTTP/s
for development and testing out-of-the-box