SWARM-616 - Repacked .war shouldn't fail on Windows (#108)
Motivation
----------

Moving the .war to .war.original before repackaging is failing on Windows.

Modifications
-------------

ShrinkWrap seems to lock the file when the archive is opened. Change the order of things so that we backup the file first, then open the backup, filter the archive and output the result back to the original location.

Result
------

Builds on Windows are fixed.