SWARM-639 - Unified meta-data for fractions/applications.
Motivation
----------

Beause of the variety of .conf files, we have a difficult time
actually keeping up with all the bootstrap artifacts, bootstrap
modules, and which dependencies are actually called for by the
application (instead of simply by WFS internally).

Modifications
-------------

All of the .conf files have been replaced in the uber-jar with
a singular META-INF/wildfly-swarm-application.yaml.  It is a yaml
file with the following useful bits of information:

- a list of bootstrap artifacts (our fraction jars as mvn coordinates)
- a list of bootstrap modules to mix into swarm.application:main, etc
- a list of actual, bonafide dependencies of the application, minus any
  transitive dependencies only required by WFS itself.

Additionally, to accomodate both the uberjar and classpath modes of executions,
similar information is held within each fraction (under META-INF/fraction-manifest.yaml)
to help make these same decisions when a META-INF/wildfly-swarm-application.yaml
is not actually present (because no uberjar).

Collectively, all of the above bits of data are glommed into a new class,
ApplicationEnvironment, which can answer the specific questions regardless
of execution mode:

* What are the app's dependencies?
* What are the bootstrap modules to add?

Result
------

In general, this should result in a no-op visible change, but
simply a cleaner, more organized internal architecture.
Also, addAllDependencies() not adding extraneous items in any
execution mode, and WEB-INF/lib/ of any maven (or gradle) -created
.war being clean of anything except application dependencies.

This is has resulted in simplifying access to all of this metadata
by wrapping it up in FractionManifest (per fraction) and
WildFlySwarmMaifest (per -swarm.jar).  These two classes are still
not used directly, but rather contribute to the user-facing singleton
of ApplicationEnvironment.

ApplicationEnvironment has also replaced Layout.