SWARM-575 - stateless-ejb javaee7 sample doesn't work
Motivation
----------

When trying to apply swarmtool commandline tool to the
stateless-ejb example, it would fail, being unable to
resolve the jta-1.0.1B artifact.  Meanwhile, the plugin
would successfully build the -swarm.jar.  Both modes
of turning a .war into a -swarm.jar should work the
same, so having a difference is bad.

Modifications
-------------

The jta-1.0.1B artifact was a symptom of a larger problem.
We were asking our ResolvingHelpers (maven, gradle and shrinkwrap)
to *transitively* resolve all the dependencies that BuildTool
figured we needed.  Which was gratuitous and ignored the fact
that the dependencies between module.xml are already handled and
solved, and did not require an additional transitive resolution
step.

Therefore, I added an additional helper.resolveAll(deps, boolean isTransitive)
that allows us to disable transitive resolution when provided with a
list of dependencies to resolve.

By default, the single-arg helper.resolveAll(deps) does work
transitively, still. The interface provides a default implementation
that calls resolveAll(deps, true) to provide the existing, legacy
behaviour of the single-parameter version.

In BuildTool's populateWhateverRepository() methods, we now
take our already-solved list of dependencies, and call resolveAll(deps, false)
to simply ensure that they are present on disk, in the uberjar
or in the user's local M2REPO.

Result
------

SwarmTool now can be applied to the stateless-ejb example without
throwing a DependencyResolutionException on jta-1.0.1B.